requirements.txt
cffi==2.0.0
cryptography==46.0.6
pillow==12.2.0
pycparser==3.0
